Beats by John
Okay, added some simple sounds that can potentially be repeated to get
some sort of tune going if we have time to do that. If there is an _
after all the characters that come before the ".sfs" type, it stands for
a question mark, as in "I'm not sure about using it give me your
thoughts". Now I'm probably headed to bed though.